# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HMS API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/hospitals/onboard": {
      "post": {
        "summary": "Public Hospital Onboarding (Self-Registration)",
        "description": "Register a new hospital along with its admin account. The admin account will be created as pending until a super admin approves it. No JWT token required.",
        "tags": [
          "Hospitals"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "hospitalName",
                  "hospitalAddress",
                  "hospitalCity",
                  "hospitalContactNumber",
                  "licenseNumber",
                  "document",
                  "adminFirstName",
                  "adminEmail",
                  "adminMobile",
                  "adminPassword"
                ],
                "properties": {
                  "hospitalName": {
                    "type": "string"
                  },
                  "hospitalAddress": {
                    "type": "string"
                  },
                  "hospitalCity": {
                    "type": "string"
                  },
                  "hospitalContactNumber": {
                    "type": "string"
                  },
                  "hospitalEmail": {
                    "type": "string"
                  },
                  "licenseNumber": {
                    "type": "string"
                  },
                  "document": {
                    "type": "string",
                    "format": "binary"
                  },
                  "adminFirstName": {
                    "type": "string"
                  },
                  "adminLastName": {
                    "type": "string"
                  },
                  "adminEmail": {
                    "type": "string"
                  },
                  "adminMobile": {
                    "type": "string"
                  },
                  "adminPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hospital and Admin registered successfully (Pending Approval)"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/v1/hospitals/create": {
      "post": {
        "summary": "Create a new hospital (Admin Only)",
        "tags": [
          "Hospitals"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "hospitalName",
                  "description",
                  "address",
                  "location",
                  "phone",
                  "email"
                ],
                "properties": {
                  "hospitalName": {
                    "type": "string",
                    "example": "City General Hospital"
                  },
                  "description": {
                    "type": "string",
                    "example": "A premier multi-specialty healthcare facility."
                  },
                  "address": {
                    "type": "object",
                    "required": [
                      "street",
                      "area",
                      "city",
                      "state",
                      "pincode"
                    ],
                    "properties": {
                      "street": {
                        "type": "string",
                        "example": "123 Health Ave"
                      },
                      "area": {
                        "type": "string",
                        "example": "Bandra West"
                      },
                      "city": {
                        "type": "string",
                        "example": "Mumbai"
                      },
                      "state": {
                        "type": "string",
                        "example": "Maharashtra"
                      },
                      "country": {
                        "type": "string",
                        "default": "India",
                        "example": "India"
                      },
                      "pincode": {
                        "type": "string",
                        "example": "400050"
                      }
                    }
                  },
                  "location": {
                    "type": "object",
                    "required": [
                      "latitude",
                      "longitude"
                    ],
                    "properties": {
                      "latitude": {
                        "type": "number",
                        "example": 19.0596
                      },
                      "longitude": {
                        "type": "number",
                        "example": 72.8295
                      }
                    }
                  },
                  "phone": {
                    "type": "string",
                    "example": "+91-9876543210"
                  },
                  "emergencyNumber": {
                    "type": "string",
                    "example": "+91-9876543211"
                  },
                  "email": {
                    "type": "string",
                    "example": "contact@cityhospital.com"
                  },
                  "services": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "title",
                        "description"
                      ],
                      "properties": {
                        "title": {
                          "type": "string",
                          "example": "Cardiology"
                        },
                        "description": {
                          "type": "string",
                          "example": "High-end cardiac care and operations."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hospital created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Not an admin)"
          }
        }
      }
    },
...
    "/api/v1/hospitals/active": {
      "get": {
        "summary": "Get details of the active hospital",
        "tags": [
          "Hospitals"
        ],
        "responses": {
          "200": {
            "description": "Hospital details retrieved successfully"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/hospitals/search": {
      "get": {
        "summary": "Search and filter hospitals",
        "tags": [
          "Hospitals"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "General text search on name, city, state, pincode, or services"
          },
          {
            "in": "query",
            "name": "city",
            "schema": {
              "type": "string"
            },
            "description": "Filter by city"
          },
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            },
            "description": "Filter by state"
          },
          {
            "in": "query",
            "name": "service",
            "schema": {
              "type": "string"
            },
            "description": "Filter by service title"
          }
        ],
        "responses": {
          "200": {
            "description": "List of hospitals matching criteria"
          }
        }
      }
    },
    "/api/v1/hospitals/nearby": {
      "get": {
        "summary": "Find nearby hospitals sorted by physical distance",
        "tags": [
          "Hospitals"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "latitude",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "User latitude"
          },
          {
            "in": "query",
            "name": "longitude",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "User longitude"
          },
          {
            "in": "query",
            "name": "maxDistanceKm",
            "schema": {
              "type": "number",
              "default": 50
            },
            "description": "Search radius in kilometers"
          }
        ],
        "responses": {
          "200": {
            "description": "Sorted list of nearby hospitals"
          },
          "400": {
            "description": "Missing coordinates"
          }
        }
      }
    },
    "/api/v1/hospitals/{id}/doctors": {
      "get": {
        "summary": "Get all doctors practicing in a specific hospital",
        "tags": [
          "Hospitals"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Hospital ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of doctors"
          }
        }
      }
    },
    "/api/v1/hospitals/{id}/upload-logo": {
      "post": {
        "summary": "Upload hospital logo image (Admin Only)",
        "tags": [
          "Hospitals"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Hospital ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "logo"
                ],
                "properties": {
                  "logo": {
                    "type": "string",
                    "format": "binary",
                    "description": "Logo image file (JPEG, JPG, PNG)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logo uploaded successfully"
          },
          "400": {
            "description": "File upload error"
          },
          "404": {
            "description": "Hospital not found"
          }
        }
      }
    },
    "/api/v1/hospitals/{id}/upload-images": {
      "post": {
        "summary": "Upload hospital infrastructure images (Admin Only)",
        "tags": [
          "Hospitals"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Hospital ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "images"
                ],
                "properties": {
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Infrastructure images (JPEG, JPG, PNG, Max 10)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Images uploaded successfully"
          },
          "400": {
            "description": "File upload error"
          },
          "404": {
            "description": "Hospital not found"
          }
        }
      }
    },
...
  },
...
}